add #import, refactor

+++ modules/builtin.c
@@ -1,12 +1,12 @@
 string :: struct
 {
-	data: *i8;
+	data: *u8;
 	count: i64;
 };
 
 CoroutineState :: struct
 {
-	state: int;
+	state: i64;
 };
 
 int64_to_float32 :: proc(value: i64) -> i32
